More cases added
This is all that I think I will be able to get done on the plane,
without online documentation. I'll do the rest tonight, after I
finish and hand in the written homework (which is already late, but
I have a few late hours left and will not suffer therefrom).